Removing some magic numbers
--HG--
branch : dev

+++ src/Orchard.Specs/Bindings/CommandLine.cs
@@ -18,7 +18,7 @@ namespace Orchard.Specs.Bindings {
                 var agent = new CommandHostAgent();
                 var input = new StringReader("");
                 var output = new StringWriter();
-                details.StatusCode = agent.RunSingleCommand(
+                details.StatusCode = (int)agent.RunSingleCommand(
                     input,
                     output,
                     "Default",
